pyproject.toml
[project]
name = "dify-workflow-monitor"
version = "0.1.0"
description = "基于 Prefect 的 Dify 工作流日志获取和报告生成系统"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Dify Team"}
]
keywords = ["dify", "workflow", "monitoring", "prefect", "logs"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "prefect>=2.14.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "loguru>=0.7.2",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
storage = [
    "oss2>=2.18.0",  # 阿里云 OSS
    "boto3>=1.29.0",  # AWS S3
]

[project.scripts]
dify-monitor = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = []

[tool.uv.sources]
# 如果需要指定特定的包源，可以在这里配置
